MAINT: restructure the backend
