+++ pkg/archiverappliance/operators_test.go
@@ -1,8 +1,11 @@
-package main
+package archiverappliance
 
 import (
 	"fmt"
 	"testing"
+
+	"github.com/sasaki77/archiverappliance-datasource/pkg/models"
+	"github.com/sasaki77/archiverappliance-datasource/pkg/testhelper"
 )
 
 func TestOperatorValidator(t *testing.T) {
@@ -18,7 +21,7 @@ func TestOperatorValidator(t *testing.T) {
 	for idx, testCase := range tests {
 		testName := fmt.Sprintf("%d: %s, %t", idx, testCase.input, testCase.output)
 		t.Run(testName, func(t *testing.T) {
-			result := OperatorValidator(testCase.input)
+			result := operatorValidator(testCase.input)
 			if testCase.output != result {
 				t.Errorf("got %v, want %v", result, testCase.output)
 			}
@@ -29,12 +32,12 @@ func TestOperatorValidator(t *testing.T) {
 func TestCreateOperatorQuery(t *testing.T) {
 	var tests = []struct {
 		name   string
-		input  ArchiverQueryModel
+		input  models.ArchiverQueryModel
 		output string
 	}{
 		{
 			name: "mean with 10 second Interval",
-			input: ArchiverQueryModel{
+			input: models.ArchiverQueryModel{
 				Operator: "mean",
 				Interval: 10,
 			},
@@ -42,7 +45,7 @@ func TestCreateOperatorQuery(t *testing.T) {
 		},
 		{
 			name: "raw with binInterval function",
-			input: ArchiverQueryModel{
+			input: models.ArchiverQueryModel{
 				Operator: "raw",
 				Interval: 16,
 			},
@@ -50,8 +53,8 @@ func TestCreateOperatorQuery(t *testing.T) {
 		},
 		{
 			name: "empty operator with 0 second interval",
-			input: ArchiverQueryModel{
-				IntervalMs: InitIntPointer(100),
+			input: models.ArchiverQueryModel{
+				IntervalMs: testhelper.InitIntPointer(100),
 				Operator:   "",
 				Interval:   0,
 			},
@@ -59,8 +62,8 @@ func TestCreateOperatorQuery(t *testing.T) {
 		},
 		{
 			name: "empty operator with 0 second interval and DisableAutoRaw true",
-			input: ArchiverQueryModel{
-				IntervalMs:     InitIntPointer(100),
+			input: models.ArchiverQueryModel{
+				IntervalMs:     testhelper.InitIntPointer(100),
 				Operator:       "",
 				Interval:       0,
 				DisableAutoRaw: true,
@@ -69,7 +72,7 @@ func TestCreateOperatorQuery(t *testing.T) {
 		},
 		{
 			name: "max operator with 10 second interval",
-			input: ArchiverQueryModel{
+			input: models.ArchiverQueryModel{
 				Operator: "max",
 				Interval: 10,
 			},
@@ -77,7 +80,7 @@ func TestCreateOperatorQuery(t *testing.T) {
 		},
 		{
 			name: "max operator with 0.1 second interval",
-			input: ArchiverQueryModel{
+			input: models.ArchiverQueryModel{
 				Operator: "max",
 				Interval: 0,
 			},
@@ -86,7 +89,7 @@ func TestCreateOperatorQuery(t *testing.T) {
 	}
 	for _, testCase := range tests {
 		t.Run(testCase.name, func(t *testing.T) {
-			result, err := CreateOperatorQuery(testCase.input)
+			result, err := createOperatorQuery(testCase.input)
 			if err != nil {
 				t.Errorf("Error received %v", err)
 			}
